requirements.txt
pybatfish >= 2021.11.4.1095
pandas >= 1.1.5
black >= 21.7b0
flake8 >= 4.0.1
pylint >= 2.13.8
jinja2 >= 2.1.1
PyYAML~=6.0